Add support for OSC "133;B" sequence
st-sx already supports the OSC "133;A" sequence, so if your shell prompt
emits that sequence, you can use the Ctrl+Shift+z/x shortcuts to scroll
to the previous prompts.

This update now adds support for the OSC "133;B" sequence, which allows
st-sx to also track where the user input begins on the shell prompt.
With this, we could also add OSC 133 support to the externalpipe patch
and the st-copyout script, so that externalpipe would add the sequences
to the output buffer and st-copyout would use them to reliably detect
where the prompts and commands begin.